+++ fabric/src/main/java/net/sashiro/compressedblocks/fabric/platform/FabricPlatformHelper.java
@@ -9,7 +9,7 @@ import net.minecraft.world.level.block.Block;
 import net.sashiro.compressedblocks.fabric.client.CompressedBlocksClient;
 import net.sashiro.compressedblocks.item.CrateItem;
 import net.sashiro.compressedblocks.platform.services.IPlatformHelper;
-import net.sashiro.compressedblocks.util.StringUtils;
+import net.sashiro.compressedblocks.util.CommonUtils;
 
 import static net.sashiro.compressedblocks.Constants.*;
 
@@ -45,7 +45,7 @@ public class FabricPlatformHelper implements IPlatformHelper {
     public void registerCrate(String name, Block... crateBlocks) {
         for (int i = 0; i < crateBlocks.length; i++) {
             FabricItemSettings fis = (FabricItemSettings) new FabricItemSettings().tab(CompressedBlocksClient.CRATE_ITEMS);
-            String prefixedName = StringUtils.getCratePrefix(i) + name;
+            String prefixedName = CommonUtils.getCratePrefix(i) + name;
             Registry.register(Registry.BLOCK, new ResourceLocation(MOD_ID, prefixedName.toLowerCase()), crateBlocks[i]);
             Registry.register(Registry.ITEM, new ResourceLocation(MOD_ID, prefixedName.toLowerCase()), new CrateItem(crateBlocks[i], fis));
             CRATES.add(crateBlocks[i]);