+++ forge/src/main/java/net/sashiro/compressedblocks/forge/data/providers/CBBlockStateProvider.java
@@ -10,7 +10,7 @@ import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.registries.RegistryObject;
 import net.sashiro.compressedblocks.block.CrateBlock;
 import net.sashiro.compressedblocks.forge.CompressedBlocksForge;
-import net.sashiro.compressedblocks.util.StringUtils;
+import net.sashiro.compressedblocks.util.CommonUtils;
 
 import static net.sashiro.compressedblocks.Constants.MOD_ID;
 
@@ -32,12 +32,12 @@ public class CBBlockStateProvider extends BlockStateProvider {
         for (RegistryObject<Block> block : CompressedBlocksForge.BLOCKS.getEntries()) {
             String descriptionId = block.get().getDescriptionId();
             // exclude manually added resources
-            if (StringUtils.isBlock(descriptionId)) continue;
+            if (CommonUtils.isBlock(descriptionId)) continue;
             // Rotational Blocks
-            if (StringUtils.isRotational(descriptionId)) {
+            if (CommonUtils.isRotational(descriptionId)) {
                 RotatedPillarBlock block1 = (RotatedPillarBlock) block.get();
-                ResourceLocation side = StringUtils.getActualResourceLocation(descriptionId);
-                ResourceLocation end = new ResourceLocation("minecraft", StringUtils.getCleanName(descriptionId) + "_top");
+                ResourceLocation side = CommonUtils.getActualResourceLocation(descriptionId);
+                ResourceLocation end = new ResourceLocation("minecraft", CommonUtils.getCleanName(descriptionId) + "_top");
 
                 if (descriptionId.contains("froglight") || descriptionId.contains("hay") || descriptionId.contains("melon") || descriptionId.contains("pumpkin")) {
                     side = new ResourceLocation("minecraft", side.getPath() + "_side");
@@ -50,7 +50,7 @@ public class CBBlockStateProvider extends BlockStateProvider {
                                         end
                                 )
                                 .texture("particle", side)
-                                .texture("overlay", StringUtils.getOverlay(descriptionId))
+                                .texture("overlay", CommonUtils.getOverlay(descriptionId))
                                 .parent(cubeColumn)
                         , models().cubeColumnHorizontal(
                                         descriptionId.replace("block.compressedblocks.", "") + "_horizontal",
@@ -58,7 +58,7 @@ public class CBBlockStateProvider extends BlockStateProvider {
                                         end
                                 )
                                 .texture("particle", side)
-                                .texture("overlay", StringUtils.getOverlay(descriptionId))
+                                .texture("overlay", CommonUtils.getOverlay(descriptionId))
                                 .parent(cubeColumnHorizontal)
                 );
             }
@@ -67,10 +67,10 @@ public class CBBlockStateProvider extends BlockStateProvider {
                 simpleBlock(
                         block.get(), models().cubeAll(
                                         descriptionId.replace("block.compressedblocks.", ""),
-                                        StringUtils.getActualResourceLocation(descriptionId)
+                                        CommonUtils.getActualResourceLocation(descriptionId)
                                 )
-                                .texture("particle", StringUtils.getActualResourceLocation(descriptionId))
-                                .texture("overlay", StringUtils.getOverlay(descriptionId))
+                                .texture("particle", CommonUtils.getActualResourceLocation(descriptionId))
+                                .texture("overlay", CommonUtils.getOverlay(descriptionId))
                                 .parent(blockBlock)
                 );
             }
@@ -79,8 +79,8 @@ public class CBBlockStateProvider extends BlockStateProvider {
         for (RegistryObject<Block> crateBlock : CompressedBlocksForge.CRATE_BLOCKS.getEntries()) {
             String descriptionId = crateBlock.get().getDescriptionId();
             String crate_name = descriptionId.replace("block.compressedblocks.", "");
-            String mc_name = StringUtils.getMCName(crate_name);
-            ResourceLocation location = StringUtils.getResourceLocation(mc_name);
+            String mc_name = CommonUtils.getMCName(crate_name);
+            ResourceLocation location = CommonUtils.getResourceLocation(mc_name);
             CrateBlock rotatedPillarBlock = (CrateBlock) crateBlock.get();
 
             horizontalBlock(
@@ -88,7 +88,7 @@ public class CBBlockStateProvider extends BlockStateProvider {
                     models().withExistingParent(crate_name, new ResourceLocation("compressedblocks", "block/template/template_cube_column_crate"))
                             .texture("all", new ResourceLocation(MOD_ID, "block/crate"))
                             .texture("item", location)
-                            .texture("number", StringUtils.getOverlay(descriptionId))
+                            .texture("number", CommonUtils.getOverlay(descriptionId))
             );
         }
     }