+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## General
 
-- changed version scheming for clarity
+- Added `Lesser Compressed Blocks`, addresses #43
+    - crafted using a 2x2 recipe instead of a 3x3 recipe
+    - for now this applies to all levels, but may be changed in the future (open an issue or post a comment if you want
+      to
+      discuss this)
 
 ---